+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 #### 2018.4.0 - Feature Release
+* Added in app user handbook
 * Added in page search
 * Added email alerts and notifications
 * Added tags in the main view
 * Added configurable default section
 * Added start over option
+* Improved import and export
 * Improved password generator
 * Improved mobile style